// src/services/types/supabase.ts
export interface definitions {
  attempts: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: string;
    created_at: string;
    updated_at: string;
    seconds: number;
    missed: number;
    words: number;
    /**
     * Note:
     * This is a Foreign Key to `sessions.id`.<fk table='sessions' column='id'/>
     */
    session_id: string;
    /**
     * Note:
     * This is a Foreign Key to `profiles.id`.<fk table='profiles' column='id'/>
     */
    user_id: string;
  };
  profiles: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: string;
    name?: string;
    avatar?: string;
  };
  profiles_private: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     * This is a Foreign Key to `profiles.id`.<fk table='profiles' column='id'/>
     */
    id: string;
    email?: string;
    admin: boolean;
  };
  sessions: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: string;
    created_at: string;
    updated_at: string;
    /**
     * Note:
     * This is a Foreign Key to `profiles.id`.<fk table='profiles' column='id'/>
     */
    user_id: string;
    lines: string;
    title: string;
  };
}